feat(stats): rewrite to temporary "plain sql" stats layer which wasn't working due to migration to webflux
More investigation has to be made to determine why this regression happen. This migration is coherent with the JOOQ (or R2DBC). Due to this migration from code to SQL the performance on this endpoint have been improve by 16x ! (99ms which was 1,6s). The Object[][] type has been used to SIMPLIFY the work on this feature due to future migration